+++ finagle-thriftmux/src/main/scala/com/twitter/finagle/thriftmux/pushsession/VanillaThriftSession.scala
@@ -63,14 +63,13 @@ private final class VanillaThriftSession(
   private[this] val locals: Local.Context =
     Local.letClear {
       val peerCertLocal =
-        handle.sslSessionInfo.peerCertificates.headOption
-          .map(Contexts.local.KeyValuePair(Transport.peerCertCtx, _)).toList
+        Contexts.local.KeyValuePair(Transport.sslSessionInfoCtx, handle.sslSessionInfo)
 
       val remoteAddressLocal =
         Contexts.local.KeyValuePair(RemoteInfo.Upstream.AddressCtx, handle.remoteAddress)
 
       Trace.letTracer(params[param.Tracer].tracer) {
-        Contexts.local.let(remoteAddressLocal :: peerCertLocal) {
+        Contexts.local.let(Seq(remoteAddressLocal, peerCertLocal)) {
           Local.save()
         }
       }